some api changes and adding support for python 3.10
