+++ vns-python/builder.py
@@ -25,7 +25,8 @@ for i in range(2):
 """
 action: str = input("Do you want to package and upload the latest build (Y/n):")
 if action == "Y":
-    Builder.upload_package("cp311")
+    Builder.build()
+    Builder.upload()
 elif action != "N":
     Builder.remove("src")
 """
