+++ vns-python/pyvns/__init__.py
@@ -1,2 +1,3 @@
-from .compiler import Compiler, Naming
+from .compiler import Compiler
 from .content import Content, ContentManager
+from .decompiler import Decompiler, Naming
